refactor(llm): sort le moteur de template de prompt dans utils/llm
prompt-template.ts est un utilitaire generique de 17 lignes, sans dependance,
qui n'a rien de specifique a l'import IA. Il s'appretait a avoir un second
consommateur hors de ai-plan-import ; le laisser sur place aurait force la
prochaine feature a importer depuis ai-plan-import, soit exactement le
couplage transverse que la remontee du dossier vient de defaire.

Le fichier et son spec rejoignent utils/llm, aux cotes de llm.service et de
parse-json-response. Signature de definePrompt et generatePrompt inchangee ;
10 imports reecrits vers l'alias @tet/backend/utils/llm/prompt-template.
