+++ apps/backend/src/plans/ai-plan-import/pipeline/score-actions/score-actions.prompt.ts
@@ -1,4 +1,4 @@
-import { generatePrompt } from '../../prompts/prompt-template';
+import { generatePrompt } from '@tet/backend/utils/llm/prompt-template';
 import { SCORING_PROMPT } from '../../prompts/scoring.prompt';
 
 export type ScoringPromptInput = {
